refactor(demo): move logs + logging interceptor into DemoModule (ADR-0001)
The access-log viewer and its global interceptor are demo concerns; the
IdP AuthModule now provides only the AuthenticatedGuard. APP_INTERCEPTOR
declared in DemoModule stays global, preserving log behavior.
